requirements.txt
# W-5 Football Prediction Framework - Dependencies

# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Gradient Boosting Models
xgboost>=2.0.0
lightgbm>=4.0.0

# LLM API Clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Data Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter Notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Type Checking (development)
mypy>=1.5.0

